snapshot of project "flist", label t19951021

+++ src/datent.h
@@ -1,4 +1,4 @@
-/* $Id: datent.h,v 1.1 1995/06/04 21:57:22 tom Exp $
+/* $Id: datent.h,v 1.2 1995/10/21 18:26:19 tom Exp $
  *
  * VMS date+time is stored as a 64-bit integer.  Use my own mode to avoid
  * long sequence of include-files for RMS.  Cover up diffs between VAX and AXP
@@ -14,8 +14,7 @@
 typedef __int64 DATENT;
 #else
 typedef struct	my_datent {
-	unsigned
-	long	date64[2];
+	unsigned date64[2];
 	} DATENT;
 #define isOkDate(p) ((p)->date64[1] != 0)
 #define isBigDate(p) ((p)->date64[1] == -1)
