+++ src/sysutils.h
@@ -1,4 +1,4 @@
-/* $Id: sysutils.h,v 1.3 1995/06/06 13:42:44 tom Exp $ */
+/* $Id: sysutils.h,v 1.4 1995/10/21 18:27:29 tom Exp $ */
 
 #ifndef SYSUTILS_H
 #define SYSUTILS_H
@@ -12,13 +12,13 @@ void	sysasctim (char *buf, DATENT *q_, int ilen);
 int	sysbintim (char *ci_, DATENT *obfr);
 
 /* sydelete.c */
-long	sysdelete (char *dspec);
+unsigned sysdelete (char *dspec);
 
 /* syfom.c */
 void	sysfom (char *co_);
 
 /* sygetmsg.c */
-void	sysgetmsg (long status, char *msg, int size_msg);
+void	sysgetmsg (unsigned status, char *msg, int size_msg);
 
 /* sygetsym.c */
 int	sysgetsym (char *co_, char *ci_, int len);
@@ -28,7 +28,7 @@ unsigned syshour (DATENT *q_);
 void	syshours (char *co_, unsigned q, int len);
 
 /* syrename.c */
-long	sysrename (char *newspec, char *oldspec);
+unsigned sysrename (char *newspec, char *oldspec);
 
 /* syrights.c */
 int	sysrights (int mask, int word_num);
